backend/pyproject.toml
[project]
name = "guided-analysis-backend"
version = "0.1.0"
description = "Guided Analysis Workflow — FastAPI backend"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "markdown>=3.7",
    "openhosta>=4.0.3",
    "pydantic-settings>=2.13.1",
    "pytrends>=4.9.2",
    "uvicorn[standard]>=0.42.0",
    "weasyprint>=62.3",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
